requirements.txt
datasets==2.12.0
matplotlib==3.7.1
nest_asyncio==1.5.6
numpy==1.23.5
opencv_python==4.7.0.72
pandas==2.0.1
Pillow==9.5.0
Pillow==9.5.0
scikit_learn==1.2.2
tensorflow==2.12.0
tensorflow_intel==2.12.0
